♻️ [refactor]: #452 Phase 6: lexer_skip_ws_and_comments.rs を分離
skip_whitespace_and_comments の 8 テストを
`src/lexer/tests/lexer_skip_ws_and_comments.rs` に分離。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_skip_ws_and_comments.rs: 59 行 (< 800)

Refs #452
